+++ src/Permission1.java
@@ -1,3 +1,4 @@
+import java.lang.reflect.InvocationTargetException;
 import java.util.UUID;
 
 /**
@@ -34,7 +35,7 @@ public final class Permission1 implements Permission {
      * @param help
      *            the description text of the permission
      */
-    private Permission1(String name, String help) {
+    public Permission1(String name, String help) {
         this.name = name;
         this.help = help;
         this.id = UUID.randomUUID().toString();
@@ -51,6 +52,11 @@ public final class Permission1 implements Permission {
         return this.id;
     }
 
+    /**
+     * main method for testing.
+     *
+     * @param args
+     */
     public static void main(String[] args) {
         Permission1 permission = new Permission1("Administrator",
                 "1. the highest permission, able to execute every command");
@@ -59,6 +65,42 @@ public final class Permission1 implements Permission {
         System.out.println("permission.getId() = " + permission.getId());
         System.out.println("permission.getName() = " + permission.getName());
 
+    }
+
+    @Override
+    public void clear() {
+        this.name = null;
+        this.help = null;
+        this.id = null;
+    }
+
+    @Override
+    public Permission newInstance() {
+        Class<?> clazz = this.getClass();
+
+        try {
+            Permission newPermission = (Permission) clazz
+                    .getDeclaredConstructor().newInstance();
+            clazz.getDeclaredField("id").set(newPermission, this.id);
+            clazz.getDeclaredField("name").set(newPermission, this.name);
+            clazz.getDeclaredField("help").set(newPermission, this.help);
+            return newPermission;
+        } catch (InstantiationException | IllegalAccessException
+                | IllegalArgumentException | NoSuchFieldException
+                | SecurityException | InvocationTargetException
+                | NoSuchMethodException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public void transferFrom(Permission arg0) {
+        assert arg0 instanceof Permission1 : ""
+                + "Violation of: arg0 is of dynamic type Permission1";
+        this.name = arg0.getName();
+        this.help = arg0.getDescription();
+        this.id = arg0.getId();
+        arg0.clear();
     }
 
 }